ML-code for RNA
Prime requirement for analysis is the .csv file of RNA data to be analysed. The RNA data can be loaded from a CSV file using pandas. The data is then split into training and test sets using train_test_split function from scikit-learn. The random forest classifier is trained using the training data and the number of estimators is set to 100. The accuracy of the model is calculated using the accuracy_score function from scikit-learn.